requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
oauthlib==3.1.1
python-decouple==3.5
requests==2.27.1
requests-oauthlib==1.3.0
tweepy==4.4.0
urllib3==1.26.8
